complete example api to get currency values from database

+++ app/main/models.py
@@ -6,7 +6,7 @@ class CurrencyExchangeRates(db.Model):
     base_currency = db.Column(db.String(20), nullable=False)
     quote_currency = db.Column(db.String(20), nullable=False)
     exchange_date = db.Column(db.Date())
-    exchange_rate = db.Column(db.Numeric(8, 8))
+    exchange_rate = db.Column(db.Float())
     source = db.Column(db.String(10), nullable=False)
 
     __table_args__ = (db.UniqueConstraint('base_currency', 'quote_currency', 'exchange_date', 'source'), )
